vecindex_finder/requirements.txt
psycopg2-binary>=2.9.3
pandas>=1.3.5
matplotlib>=3.5.1
seaborn>=0.11.2
pyyaml>=6.0
tqdm>=4.62.3
numpy>=1.21.5 
pydantic>=2.10.6